TradeBitCoin/requirements.txt
# requirements.txt
python-dotenv
ccxt
pandas
anthropic
google-generativeai 